+++ core/db/query-types/class-wc-products.php
@@ -36,6 +36,8 @@ class WC_Products extends WP_Query_Base {
 		) );
 		$query_args['include_variations'] = wponion_is_bool_val( $query_args['include_variations'] );
 		$query_args['display_stock']      = wponion_is_bool_val( $query_args['display_stock'] );
+		$query_args['include']            = ( isset( $query_args['post__in'] ) && ! empty( $query_args['post__in'] ) ) ? $query_args['post__in'] : $query_args['include'];
+		$query_args['exclude']            = ( isset( $query_args['post__not_in'] ) && ! empty( $query_args['post__not_in'] ) ) ? $query_args['post__not_in'] : $query_args['exclude'];
 		return $query_args;
 	}