+++ data/fonts/google.php
@@ -1,5 +1,5 @@
 <?php
-// Last Updated : 2020-06-14 05:49:33:am
+// Last Updated : 2020-06-17 05:49:16:am
 if ( ! defined( "ABSPATH") ) { die; } 
 return array (
   'ABeeZee' =>