+++ wponion-init.php
@@ -2,8 +2,8 @@
 
 defined( 'ABSPATH' ) || exit;
 
-defined( 'WPONION_VERSION' ) or define( 'WPONION_VERSION', '1.5.2' );
-defined( 'WPONION_CDN_VERSION' ) or define( 'WPONION_CDN_VERSION', '1.3.9' );
+defined( 'WPONION_VERSION' ) or define( 'WPONION_VERSION', '1.5.3' );
+defined( 'WPONION_CDN_VERSION' ) or define( 'WPONION_CDN_VERSION', '1.4' );
 defined( 'WPONION_CDN_URL' ) or define( 'WPONION_CDN_URL', 'https://cdn.jsdelivr.net/gh/wponion/vendor-support' );
 defined( 'WPONION_NAME' ) or define( 'WPONION_NAME', 'WPOnion' );
 defined( 'WPONION_PATH' ) or define( 'WPONION_PATH', plugin_dir_path( __FILE__ ) );